Conform to PEP8
We can't fix some things:

config.sample.py:25:3: W291 trailing whitespace
This is the signature separator that is defined as "-- " with the
trailing space.

NotifierManager.py:68:68: E711 comparison to None should be 'if cond is None:'
Those comparisons aren't really Python comparisons, but are interpreted
by SQLAlchemy, and aren't understood if "is None" is used instead.